requirements.txt
# requirements.txt
pandas==2.1.0
scikit-learn==1.3.0
openai==1.12.0
azure-storage-blob==12.19.0
azure-storage-queue==12.9.0
azure-monitor-opentelemetry==1.2.0
black==23.7.0
pytest==7.4.0